Clarify project licensing. (#1)
* Clarify license information.

Add LICENSE file holding the Polyform Strict License 1.0.0.
Update the README to reference this license and explain the possibility
of changing the license in the future.

Some information about the Polyform Strict License 1.0.0:

Recipients of the software (the licensees) are not permitted to change,
distribute, or create new works based on the software. (See "Copyright
License")

Use of the software is permitted for a limited set of uses and
organizations.

The uses importantly include "research, experiment, and testing for the
benefit of public knowledge, personal study". (See "Personal Uses")

The types of organizations include "charitable organization,
educational institution, public research organization". (See
"Noncommerical Organizations")

Licensees are not permitted to relicense or transfer licenses to anyone
else. (See "No other rights")

Noncompliance with the terms of this license must be rectified within 32
days of recieving notice from the licensor. (See "Violations")

* Update copyright statement in each Python file.

Update the copyright statement in each Python file to fully spell out
"World Resources Institute" rather than using "WRI".

* Update license in source files to Polyform Strict.